realtek: Add support for SFP EEPROM-access over SMBus
The EEPROMs on SFP modules are compatible both to I2C as well
as SMBus. However, the kernel so far only supports I2C
access. We add SMBus access routines, because the I2C driver
for the RTL9300 HW only supports that protocol. At the same
time we disable I2C access to PHYs on SFP modules as otherwise
detection of any SFP module would fail. This is not in any
way problematic at this point in time since the RTL93XX
platform so far does not support PHYs on SFP modules. An
alternative is to implement PHY over SMBus access, see e.g.
the link below.

The patch is inspired by the work here:
https://bootlin.com/blog/sfp-modules-on-a-board-running-linux/

Signed-off-by: Birger Koblitz <git@birger-koblitz.de>